Make useable as module >>> from whatportis import get_ports >>> get_ports("mysql") [Port(name=mysql, port=3306, protocol=tcp, description=MySQL), Port(name=mysql, port=3306, protocol=udp, description=MySQL)] >>>
