+++ whatportis/__init__.py
@@ -1,1 +1,5 @@
-from whatportis.main import get_ports
+# -*- coding: utf-8 -*-
+
+from .core import get_ports
+
+__all__ = ["get_ports"]
